VMValue: Uniform, typed variable registry introduced
The new VMValue structure encapsulates all value types (Int64, Double, Bool, String) for the variable registry and replaces the previous tuples. IVariableRegistry and its implementation have been adapted accordingly and extended with typed access methods.
All relevant components (including EvaluateCommand, Lowering, ExpressionEvaluator) now use VMValue for variable access and replacement.
ScriptCommand and ScriptStepperExtension receive the registry in the constructor and pass it on to WeaverProgram. This makes the typing and handling of variables more consistent, secure, and extensible throughout the scripting system. Tests and callers have been adjusted accordingly.

Still fighting with Runtime Variable replacement. At least I found the core Issue in ExpressionEvaluator the first and the last place we must replace it

+++ Mediator/Commands/RegistryCommandTests.cs
@@ -57,7 +57,7 @@ namespace Mediator.Commands
 
             var getResult = _getValue.Execute("counter");
             Assert.IsTrue(getResult.Success);
-            Assert.AreEqual(42, getResult.Value);
+            Assert.AreEqual((long)42, getResult.Value);
             Assert.AreEqual(EnumTypes.Wint, getResult.Type);
         }
